remove fairing_util and move the remaining to util instead
