+++ xgboost_synthetic/util.py
@@ -1,6 +1,10 @@
 import os
 import shutil
 import subprocess
+import json
+import numpy as np
+import requests
+from retrying import retry
 
 KFP_PACKAGE = 'https://storage.googleapis.com/ml-pipeline/release/0.1.20/kfp.tar.gz'
 def notebook_setup():
@@ -25,3 +29,19 @@ def copy_data_to_nfs(nfs_path, model_dir):
 
     if not os.path.exists(model_dir):
         os.makedirs(model_dir)
+
+@retry(wait_exponential_multiplier=1000, wait_exponential_max=5000,
+       stop_max_delay=2*60*1000)
+def predict_nparray(url, data, feature_names=None):
+    pdata={
+        "data": {
+            "names":feature_names,
+            "tensor": {
+                "shape": np.asarray(data.shape).tolist(),
+                "values": data.flatten().tolist(),
+            },
+        }
+    }
+    serialized_data = json.dumps(pdata)
+    r = requests.post(url, data={'json':serialized_data}, timeout=5)
+    return r
